Fix a warning when copying the webapi port token int into its destination ushort.
git-svn-id: svn://svn.icculus.org/fs2open/trunk/fs2_open@10232 387891d4-d844-0410-90c0-e4c51a9137d3

+++ code/network/multi_options.cpp
@@ -199,7 +199,7 @@ void multi_options_read_config()
 				if ( SETTING("+webapi_server_port") ) {
 					NEXT_TOKEN();
 					if (tok != NULL) {
-						Multi_options_g.webapiPort = atoi(tok);
+						Multi_options_g.webapiPort = (ushort) atoi(tok);
 					}
 				}
 			}
